JournalService refactor: from service to singleton.
