+++ EasyTargetDiet/src/main/java/br/com/arndroid/etdiet/provider/days/DaysManager.java
@@ -4,6 +4,7 @@ import android.content.ContentUris;
 import android.content.Context;
 import android.database.Cursor;
 import android.net.Uri;
+import android.util.Log;
 
 import java.util.Calendar;
 import java.util.Date;
@@ -65,12 +66,26 @@ public class DaysManager {
 
         final Long id = entity.getId();
         if(id == null) {
+            if (isLogEnabled) {
+                Log.d(TAG, "About to insert entity=" + entity + "with Uri=" + Contract.Days.CONTENT_URI);
+            }
             final Uri resultUri = mContext.getContentResolver().insert(Contract.Days.CONTENT_URI,
                     entity.toContentValues());
             entity.setId(Long.parseLong(resultUri.getLastPathSegment()));
+            if (isLogEnabled) {
+                Log.d(TAG, "Returning from insert: entity inserted with id=" + entity.getId() + " and dateId=" + entity.getDateId());
+            }
         } else {
             mContext.getContentResolver().update(ContentUris.withAppendedId(Contract.Days.CONTENT_URI,id),
                     entity.toContentValues(), null, null);
+            if (isLogEnabled) {
+                Log.d(TAG, "Returning from update: entity updated =" + entity);
+            }
         }
     }
+
+    @SuppressWarnings("UnusedDeclaration")
+    private static final String TAG = "==>ETD/" + DaysManager.class.getSimpleName();
+    @SuppressWarnings("UnusedDeclaration")
+    private static final boolean isLogEnabled = true;
 }
