+++ EasyTargetDiet/src/main/java/br/com/arndroid/etdiet/util/sk/m217/tests/utils/ProviderTestCase3.java
@@ -1,4 +1,4 @@
-package sk.m217.tests.utils;
+package br.com.arndroid.etdiet.util.sk.m217.tests.utils;
 /*
  * Copyright (C) 2011 The Android Open Source Project
  *
