+++ EasyTargetDiet/src/main/java/br/com/arndroid/etdiet/util/sk/m217/tests/utils/readme.txt
