--- EasyTargetDiet/src/main/java/br/com/arndroid/etdiet/virtualweek/JournalApi.java
@@ -1,10 +1,0 @@
-package br.com.arndroid.etdiet.virtualweek;
-
-public interface JournalApi {
-
-    public void registerViewObserver(ViewObserver observer);
-
-    public void unregisterViewObserver(ViewObserver observer);
-
-    public void requestSummaryForDateId(ViewObserver observer, String dateId);
-}
