+++ EasyTargetDiet/src/main/java/br/com/arndroid/etdiet/virtualweek/VirtualWeek.java
@@ -1,222 +1,234 @@
 package br.com.arndroid.etdiet.virtualweek;
 
+import android.content.ContentResolver;
+import android.content.ContentUris;
 import android.content.Context;
+import android.net.Uri;
 
-import java.util.Calendar;
 import java.util.Date;
+import java.util.LinkedList;
+import java.util.List;
 
 import br.com.arndroid.etdiet.provider.Contract;
-import br.com.arndroid.etdiet.provider.days.DaysEntity;
-import br.com.arndroid.etdiet.provider.days.DaysManager;
-import br.com.arndroid.etdiet.provider.foodsusage.FoodsUsageManager;
-import br.com.arndroid.etdiet.provider.parametershistory.ParametersHistoryManager;
-import br.com.arndroid.etdiet.util.DateUtil;
+import br.com.arndroid.etdiet.util.sk.m217.tests.utils.MockContentResolver2;
 
-public class VirtualWeek {
+public class VirtualWeek implements DatabaseChangeObserver.ChangeListener {
 
-    public static final int DAYS_IN_A_WEEK = 7;
-    public static final int LAST_WEEKDAY_INDEX = DAYS_IN_A_WEEK - 1;
-    public static final int FIRST_WEEKDAY_INDEX = 0;
-    public static final int NO_INDEX = -1;
-    private static final float FLOAT_ZERO = 0.0f;
-    private static final float FORCE_CHAIN_CALC = -1.0f;
-    private Context mContext;
-    private WeekPeriod mWeekPeriod;
-    private DaySummary[] mDaySummaryArray = new DaySummary[DAYS_IN_A_WEEK];
-    private float mInitialWeeklyAllowance;
-    private int mExerciseUseMode;
-    private int mExerciseUseOrder;
+    private MockContentResolver2 mMockContentResolver2;
+    private static VirtualWeek mInstance = null;
+    private final Context mContext;
+    private VirtualWeekEngine mVirtualWeekEngine;
+    private DatabaseChangeObserver[] mDayObserverArray =
+            new DatabaseChangeObserver[VirtualWeekEngine.DAYS_IN_A_WEEK];
+    private DatabaseChangeObserver[] mFoodUsageObserverArray =
+            new DatabaseChangeObserver[VirtualWeekEngine.DAYS_IN_A_WEEK];
+    private DatabaseChangeObserver mParametersHistoryObserver = null;
+    private DatabaseChangeObserver mWeekdayParametersObserver = null;
+    private List<ViewObserver> viewObservers = new LinkedList<ViewObserver>();
 
-    public VirtualWeek(Context context, Date referenceDate) {
+    public static VirtualWeek getInstance(Context context) {
+        /*
+         We don't worry about lost of one or two instances in a multi-threaded scenario (which will
+         be very rare).
+         */
+        if (mInstance == null) {
+            mInstance = new VirtualWeek(context, null);
+        }
+        return mInstance;
+    }
+
+    public static VirtualWeek getInstanceForTests(Context context, MockContentResolver2 mockContentResolver2) {
+        return new VirtualWeek(context, mockContentResolver2);
+    }
+
+    protected VirtualWeek(Context context, MockContentResolver2 mockContentResolver2) {
         mContext = context;
-
-        for(int i = FIRST_WEEKDAY_INDEX; i < DAYS_IN_A_WEEK; i++) {
-            mDaySummaryArray[i] = new DaySummary();
-        }
-
-        buildAllDaysForReferenceDate(referenceDate);
-        calculateAllDays();
+        mMockContentResolver2 = mockContentResolver2;
+        createVirtualWeekAndRegisterObservers();
     }
 
-    private void calculateAllDays() {
-
-        for(int i = FIRST_WEEKDAY_INDEX; i < DAYS_IN_A_WEEK; i++) {
-            mDaySummaryArray[i].setWeeklyAllowanceAfterUsage(FORCE_CHAIN_CALC);
-            mDaySummaryArray[i].setExerciseToCarry(FORCE_CHAIN_CALC);
-        }
-        calculateDayAndNextIfNecessary(FIRST_WEEKDAY_INDEX);
+    private void createVirtualWeekAndRegisterObservers() {
+        mVirtualWeekEngine = new VirtualWeekEngine(mContext, new Date());
+        registerObservers();
     }
 
-    public void calculateDayAndNextIfNecessary(int index) {
-        final DaySummary currentSummary = mDaySummaryArray[index];
+    @Override
+    protected void finalize() throws Throwable {
+        try {
+            /*
+            We are calling this here because Android doesn't have a hook to intercept global app
+            shutdown.
+             */
+            unregisterObservers();
+        } finally {
+            super.finalize();
+        }
+    }
 
-        currentSummary.setUsage(new FoodsUsageManager(mContext).usageSummaryForDateId(
-                currentSummary.getEntity().getDateId()));
-        // Making an alias, but correct and updated:
-        final UsageSummary currentUsage = currentSummary.getUsage();
+    private void registerObservers() {
+        for(int i = VirtualWeekEngine.FIRST_WEEKDAY_INDEX; i < VirtualWeekEngine.DAYS_IN_A_WEEK; i++) {
+            createAndRegisterDayObserverForIndex(i);
+            createAndRegisterFoodUsageObserverForIndex(i);
+        }
+        createAndRegisterParametersObservers();
+    }
 
-        float oldWeeklyAllowanceAfterUsage = currentSummary.getWeeklyAllowanceAfterUsage();
-        float oldExerciseToCarry = currentSummary.getExerciseToCarry();
+    private void createAndRegisterParametersObservers() {
+        ContentResolver resolver = mContext.getContentResolver();
 
+        mParametersHistoryObserver = new DatabaseChangeObserver(DatabaseChangeObserver.PARAMETERS_HISTORY_CHANGE_TYPE,
+                VirtualWeekEngine.NO_INDEX, this);
+        resolver.registerContentObserver(Contract.ParametersHistory.CONTENT_URI, true,
+                mParametersHistoryObserver);
+        mWeekdayParametersObserver = new DatabaseChangeObserver(DatabaseChangeObserver.WEEKDAY_PARAMETERS_CHANGE_TYPE,
+                VirtualWeekEngine.NO_INDEX, this);
+        resolver.registerContentObserver(Contract.WeekdayParameters.CONTENT_URI, true, mWeekdayParametersObserver);
+        if (mMockContentResolver2 != null) {
+            mMockContentResolver2.safeRegisterContentObserver(Contract.ParametersHistory.CONTENT_URI,
+                    true, mParametersHistoryObserver);
+            mMockContentResolver2.safeRegisterContentObserver(Contract.WeekdayParameters.CONTENT_URI,
+                    true, mWeekdayParametersObserver);
+        }
+    }
 
-        if(index == FIRST_WEEKDAY_INDEX) {
-            currentSummary.setWeeklyAllowanceBeforeUsage(mInitialWeeklyAllowance);
-            currentSummary.setTotalExercise(currentUsage.getExerciseDone());
+    private void createAndRegisterFoodUsageObserverForIndex(int index) {
+        mFoodUsageObserverArray[index] = new DatabaseChangeObserver(DatabaseChangeObserver.FOODS_USAGE_CHANGE_TYPE,
+                index, this);
+        final Uri foodUri = Uri.withAppendedPath(Contract.FoodsUsage.DATE_ID_CONTENT_URI,
+                mVirtualWeekEngine.getDateIdForIndex(index));
+        mContext.getContentResolver().registerContentObserver(foodUri, true, mFoodUsageObserverArray[index]);
+        if (mMockContentResolver2 != null) {
+            mMockContentResolver2.safeRegisterContentObserver(foodUri, true, mFoodUsageObserverArray[index]);
+        }
+    }
+
+    private void createAndRegisterDayObserverForIndex(int index) {
+        mDayObserverArray[index] = new DatabaseChangeObserver(DatabaseChangeObserver.DAY_CHANGE_TYPE,
+                index, this);
+        final Uri dayUri;
+        if(mVirtualWeekEngine.getIdForIndex(index) == null) {
+            dayUri = Uri.withAppendedPath(Contract.Days.DATE_ID_CONTENT_URI,
+                    mVirtualWeekEngine.getDateIdForIndex(index));
         } else {
-            currentSummary.setWeeklyAllowanceBeforeUsage(mDaySummaryArray[index - 1].getWeeklyAllowanceAfterUsage());
-            currentSummary.setTotalExercise(mDaySummaryArray[index - 1].getExerciseToCarry()
-                    + currentUsage.getExerciseDone());
+            dayUri = ContentUris.withAppendedId(Contract.Days.CONTENT_URI,
+                    mVirtualWeekEngine.getIdForIndex(index));
+        }
+        mContext.getContentResolver().registerContentObserver(dayUri, true, mDayObserverArray[index]);
+        if (mMockContentResolver2 != null) {
+            mMockContentResolver2.safeRegisterContentObserver(dayUri, true, mDayObserverArray[index]);
+        }
+    }
 
+    private void unregisterObservers() {
+        for(int i = VirtualWeekEngine.FIRST_WEEKDAY_INDEX; i < VirtualWeekEngine.DAYS_IN_A_WEEK; i++) {
+            unregisterAndReleaseDayObserverForIndex(i);
+            unregisterAndReleaseFoodUsageObserverForIndex(i);
+        }
+        unregisterAndReleaseParametersObservers();
+    }
+
+    private void unregisterAndReleaseParametersObservers() {
+        if (mMockContentResolver2 != null) {
+            mMockContentResolver2.safeUnregisterContentObserver(mParametersHistoryObserver);
+            mMockContentResolver2.safeUnregisterContentObserver(mWeekdayParametersObserver);
+        }
+        ContentResolver resolver = mContext.getContentResolver();
+        resolver.unregisterContentObserver(mParametersHistoryObserver);
+        mParametersHistoryObserver = null;
+        resolver.unregisterContentObserver(mWeekdayParametersObserver);
+        mWeekdayParametersObserver = null;
+    }
+
+    private void unregisterAndReleaseFoodUsageObserverForIndex(int index) {
+        if (mMockContentResolver2 != null) {
+            mMockContentResolver2.safeUnregisterContentObserver(mFoodUsageObserverArray[index]);
+        }
+        mContext.getContentResolver().unregisterContentObserver(mFoodUsageObserverArray[index]);
+        mFoodUsageObserverArray[index] = null;
+    }
+
+    private void unregisterAndReleaseDayObserverForIndex(int index) {
+        if (mMockContentResolver2 != null) {
+            mMockContentResolver2.safeUnregisterContentObserver(mDayObserverArray[index]);
+        }
+        mContext.getContentResolver().unregisterContentObserver(mDayObserverArray[index]);
+        mDayObserverArray[index] = null;
+    }
+
+    @Override
+    public void onContentChanged(int changeType, int index) {
+        switch (changeType) {
+            case DatabaseChangeObserver.DAY_CHANGE_TYPE:
+                boolean modifyObserver = mVirtualWeekEngine.getIdForIndex(index) == null;
+                mVirtualWeekEngine.rebuildAndCalculateDayForIndex(index);
+                if(modifyObserver && mDayObserverArray[index] != null) {
+                    unregisterAndReleaseDayObserverForIndex(index);
+                    createAndRegisterDayObserverForIndex(index);
+                }
+                break;
+            case DatabaseChangeObserver.FOODS_USAGE_CHANGE_TYPE:
+                mVirtualWeekEngine.calculateDayAndNextIfNecessary(index);
+                break;
+            case DatabaseChangeObserver.PARAMETERS_HISTORY_CHANGE_TYPE:
+            case DatabaseChangeObserver.WEEKDAY_PARAMETERS_CHANGE_TYPE:
+                unregisterObservers();
+                createVirtualWeekAndRegisterObservers();
+                break;
+            default:
+                throw new IllegalStateException("Invalid changeType = " + changeType);
         }
 
-        currentSummary.setDiaryAllowanceAfterUsage(currentSummary.getEntity().getAllowed()
-                - currentUsage.getTotalUsed());
+        notifyAllViewObservers(changeType, index);
+    }
 
-        boolean exerciseAfterUsageComputed = false;
-        boolean weeklyAllowanceAfterUsageComputed = false;
-
-        if(currentSummary.getDiaryAllowanceAfterUsage() < FLOAT_ZERO
-                && mExerciseUseMode != Contract.ParametersHistory.DONT_USE_EXERCISE_USE_MODE
-                && mExerciseUseOrder == Contract.ParametersHistory.EXERCISES_FIRST_EXERCISE_USE_ORDER) {
-            currentSummary.setDiaryAllowanceAfterUsage(
-                    currentSummary.getDiaryAllowanceAfterUsage()
-                            + currentSummary.getTotalExercise());
-            currentSummary.setExerciseAfterUsage(FLOAT_ZERO);
-            exerciseAfterUsageComputed = true;
-            if(currentSummary.getDiaryAllowanceAfterUsage() > FLOAT_ZERO) {
-                currentSummary.setExerciseAfterUsage(
-                        currentSummary.getDiaryAllowanceAfterUsage());
-                currentSummary.setDiaryAllowanceAfterUsage(FLOAT_ZERO);
+    private void notifyAllViewObservers(int changeType, int index) {
+        DaySummary daySummary;
+        for (ViewObserver viewObserver : viewObservers) {
+            switch (changeType) {
+                case DatabaseChangeObserver.DAY_CHANGE_TYPE:
+                    daySummary = mVirtualWeekEngine.daySummaryForIndex(index);
+                    viewObserver.onDayChanged(daySummary);
+                    break;
+                case DatabaseChangeObserver.FOODS_USAGE_CHANGE_TYPE:
+                    daySummary = mVirtualWeekEngine.daySummaryForIndex(index);
+                    viewObserver.onFoodsUsageChanged(daySummary);
+                    break;
+                case DatabaseChangeObserver.PARAMETERS_HISTORY_CHANGE_TYPE:
+                case DatabaseChangeObserver.WEEKDAY_PARAMETERS_CHANGE_TYPE:
+                    viewObserver.onParametersChanged();
+                    break;
+                default:
+                    throw new IllegalStateException("Invalid changeType = " + changeType);
             }
         }
-
-        if(currentSummary.getDiaryAllowanceAfterUsage() < FLOAT_ZERO) {
-            currentSummary.setDiaryAllowanceAfterUsage(
-                    currentSummary.getDiaryAllowanceAfterUsage()
-                            + currentSummary.getWeeklyAllowanceBeforeUsage());
-            currentSummary.setWeeklyAllowanceAfterUsage(FLOAT_ZERO);
-            weeklyAllowanceAfterUsageComputed = true;
-            if(currentSummary.getDiaryAllowanceAfterUsage() > FLOAT_ZERO) {
-                currentSummary.setWeeklyAllowanceAfterUsage(currentSummary.getDiaryAllowanceAfterUsage());
-                currentSummary.setDiaryAllowanceAfterUsage(FLOAT_ZERO);
-            }
-        }
-
-        if(currentSummary.getDiaryAllowanceAfterUsage() < FLOAT_ZERO
-                && mExerciseUseMode != Contract.ParametersHistory.DONT_USE_EXERCISE_USE_MODE
-                && mExerciseUseOrder == Contract.ParametersHistory.WEEKLY_FIRST_EXERCISE_USE_ORDER) {
-            currentSummary.setDiaryAllowanceAfterUsage(currentSummary.getDiaryAllowanceAfterUsage()
-                + currentSummary.getTotalExercise());
-            currentSummary.setExerciseAfterUsage(FLOAT_ZERO);
-            exerciseAfterUsageComputed = true;
-            if(currentSummary.getDiaryAllowanceAfterUsage() > FLOAT_ZERO) {
-                currentSummary.setExerciseAfterUsage(currentSummary.getDiaryAllowanceAfterUsage());
-                currentSummary.setDiaryAllowanceAfterUsage(FLOAT_ZERO);
-            }
-        }
-
-        if(!weeklyAllowanceAfterUsageComputed) {
-            currentSummary.setWeeklyAllowanceAfterUsage(currentSummary.getWeeklyAllowanceBeforeUsage());
-        }
-
-        if(!exerciseAfterUsageComputed) {
-            currentSummary.setExerciseAfterUsage(currentSummary.getTotalExercise());
-        }
-
-        if(mExerciseUseMode == Contract.ParametersHistory.USE_AND_ACCUMULATE_EXERCISE_USE_MODE) {
-            currentSummary.setExerciseToCarry(currentSummary.getExerciseAfterUsage());
-        } else {
-            currentSummary.setExerciseToCarry(FLOAT_ZERO);
-        }
-
-        if(oldWeeklyAllowanceAfterUsage != currentSummary.getWeeklyAllowanceAfterUsage()
-                || oldExerciseToCarry != currentSummary.getExerciseToCarry()) {
-            if(index != LAST_WEEKDAY_INDEX) calculateDayAndNextIfNecessary(index + 1);
-        }
     }
 
-    private void buildAllDaysForReferenceDate(Date referenceDate) {
-
-        mWeekPeriod = new WeekPeriod(mContext, referenceDate);
-
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTime(mWeekPeriod.getInitialDate());
-        for(int i = FIRST_WEEKDAY_INDEX; i < DAYS_IN_A_WEEK; i++) {
-            mDaySummaryArray[i].setEntity(buildDayForDate(calendar.getTime()));
-            calendar.add(Calendar.DAY_OF_MONTH, 1);
-        }
-
-        ParametersHistoryManager parametersHistoryManager = new ParametersHistoryManager(mContext);
-        Date lastWeekday = DateUtil.dateIdToDate(mDaySummaryArray[LAST_WEEKDAY_INDEX].getEntity().getDateId());
-        mInitialWeeklyAllowance = parametersHistoryManager.getWeeklyAllowanceForDate(
-                lastWeekday);
-        mExerciseUseMode = parametersHistoryManager.getExerciseUseModeForDate(lastWeekday);
-        mExerciseUseOrder = parametersHistoryManager.getExerciseUseOrderForDate(lastWeekday);
+    public void registerViewObserver(ViewObserver observer) {
+        viewObservers.add(observer);
     }
 
-    private DaysEntity buildDayForDate(Date date) {
-        return new DaysManager(mContext).dayFromDate(date);
+    public void unregisterViewObserver(ViewObserver observer) {
+        viewObservers.remove(observer);
     }
 
-    public void rebuildAndCalculateDayForIndex(int index) {
-        final Date date = DateUtil.dateIdToDate(mDaySummaryArray[index].getEntity().getDateId());
-        mDaySummaryArray[index].setEntity(buildDayForDate(date));
-        calculateDayAndNextIfNecessary(index);
+    public void requestSummaryForDateId(ViewObserver observer, String dateId) {
+        observer.onSummaryRequested(mVirtualWeekEngine.daySummaryForDateId(dateId));
     }
 
-    public Long getIdForIndex(int index) {
-        return mDaySummaryArray[index].getEntity().getId();
+    public static interface ViewObserver {
+
+        public void onDayChanged(DaySummary summary);
+
+        public void onFoodsUsageChanged(DaySummary summary);
+
+        public void onParametersChanged();
+
+        public void onSummaryRequested(DaySummary daySummary);
     }
 
-    public String getDateIdForIndex(int index) {
-        return mDaySummaryArray[index].getEntity().getDateId();
-    }
 
-    public float getTotalUsedForIndex(int index) {
-        return mDaySummaryArray[index].getUsage().getTotalUsed();
-    }
-
-    public float getExerciseAfterUsageForIndex(int index) {
-        return mDaySummaryArray[index].getExerciseAfterUsage();
-    }
-
-    public float getAllowedForIndex(int index) {
-        return mDaySummaryArray[index].getEntity().getAllowed();
-    }
-
-    public float getWeeklyAllowanceBeforeUsageForIndex(int index) {
-        return mDaySummaryArray[index].getWeeklyAllowanceBeforeUsage();
-    }
-
-    public float getTotalExerciseForIndex(int index) {
-        return mDaySummaryArray[index].getTotalExercise();
-    }
-
-    public float getWeeklyAllowanceAfterUsageForIndex(int index) {
-        return mDaySummaryArray[index].getWeeklyAllowanceAfterUsage();
-    }
-
-    public float getDiaryAllowanceAfterUsageForIndex(int index) {
-        return mDaySummaryArray[index].getDiaryAllowanceAfterUsage();
-    }
-
-    public DaySummary daySummaryForIndex(int index) {
-        return new DaySummary(mDaySummaryArray[index]);
-    }
-
-    public DaySummary daySummaryForDateId(String dateId) {
-
-        final int index = indexForDateId(dateId);
-        return index == NO_INDEX ? null : mDaySummaryArray[index];
-    }
-
-    private int indexForDateId(String dateId) {
-
-        for (int i = 0; i < mDaySummaryArray.length; i++) {
-            if (mDaySummaryArray[i].getEntity().getDateId().equals(dateId)) {
-                return i;
-            }
-        }
-        return NO_INDEX;
-    }
+    @SuppressWarnings("UnusedDeclaration")
+    private static final String TAG = "==>ETD/" + VirtualWeek.class.getSimpleName();
+    @SuppressWarnings("UnusedDeclaration")
+    private static final boolean isLogEnabled = true;
 }
