+++ EasyTargetDiet/src/main/res/layout/journal_act.xml
@@ -45,12 +45,4 @@
         android:layout_gravity="center"
         android:onClick="cmdTestBreakfast" />
 
-    <Button
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Test Observer!"
-        android:id="@+id/btnTestService"
-        android:layout_gravity="center"
-        android:onClick="cmdTestObserver" />
-
 </LinearLayout>
